Skip caching data into Alluxio during transformation
For transformation jobs on structured table, the output should not be
written to Alluxio memory space, which might evict other active dataset,
instead, write directly through to the under storage. Also, when reading
the source files, it should be read with NO_CACHE so that the big tables
are not cached into Alluxio.

pr-link: Alluxio/alluxio#10177
change-id: cid-42cfb601d25a9f12bc13d648f9397d3566b13c79