Load metadata for the path that we are syncing
This fixes the issue #8531 .

Previously, we do not add the path we are loading to the
`UfsSyncPathCache`. So when only children need to be synced, we
repeatedly request children information from the UFS but never call
`loadMetadata` on the parent and therefore the parent is never in the
`UfsSyncPathCache`, leaving the impression that we are syncing the
parent repeatedly.

Now with this change, we do sync the parent and add it to the
`UfsSyncPathCache`, so we do not repeatedly make requests to the ufs.

pr-link: Alluxio/alluxio#8559
change-id: cid-b973b5ad7dbbc58daaa82a165852d6293df1e4a8