Set port reuse for web server
This helps with the flakiness for testing, since we set the port reuse
option for other server sockets. In unittests, a server socket is
created, the port is retrieved ("reserved"), then the socket is closed,
and then the actual server is started with that port. Port reuse is
important for the actual server to start with that same port.

pr-link: Alluxio/alluxio#10441
change-id: cid-bf7b000996de2b9c31f543ba9bd62caa8f131c1a