Propagate UserState to PollingMasterInquireClient
This is a manual back port of
https://github.com/Alluxio/alluxio/pull/10128 to `branch-2.0`.
Currently PollingMasterInquireClient uses a new Subject when querying
master. If the call needs authentication using credentials from Subject,
it will fail the request because the credentials are missing. This
change add parameters to APIs to allow passing in the required Subject.

pr-link: Alluxio/alluxio#10478
change-id: cid-91e25b9c0ad830e95a0a1893197ab70a890db2fe