Include integration test results in code coverage
Jacoco generates coverage data in .exec files, which are then used for
the jacoco-report goal. However, integration tests are in a separate
module from the source code they are testing. Therefore, all the
coverage data is not included when reporting on the source code.

Therefore, to capture all the test coverage, the reporting has to be
done in 2 steps:
- generate all coverage data (into the same data file)
- report the coverage data using the coverage data which includes all
the tests

This is done with 2 separate profiles for the different steps:
- `jacoco` to generate the coverage data
- `jacocoReport` to generate the report using the coverage data

pr-link: Alluxio/alluxio#10555
change-id: cid-7e528830d8e102c873cd9ff6a0ab3c4332f9f7fc