Fix Command Execution order on workers
Previously, the order of commands was not honored by the workers because
there were 4 threads handling the command handling service. But the
benefit of these threads is minimal because underlying
TaskExecutorManager is synchronized.

Resolve this by moving the expensive portion that could be executed in
parallel under TaskExecutor and making CommandHandlingService single
threaded.

pr-link: Alluxio/alluxio#10654
change-id: cid-2ed274b830d91c297308186f761321f2e6226597