Keep worker client open after creating data reader/writer
This fix keeps the BlockWorkerClient open after returning the
GrpcDataWriter.
Previously, we used a try-with-resources which would close the client
before
being returned to the user. This could result in poor behavior if the
writer
is used after the #create() call.

Also, to prevent a leak in the LocalFileDataReader, we attempt to close
the
reader within the try block inside of its close method.

This is a follow-up to #10719

pr-link: Alluxio/alluxio#10728
change-id: cid-93141241c264da235b7a3d73b34dde861213b062