Use one StatefulSet for multiple masters
This change covers several Helm tech debts:
**1. Use single StatefulSet for all masters. Create one NodePort Service
for each Pod.**

**2. Use PV for embedded journal instead of `emptyDir`.**
This will require one PV+PVC pair to work for each master Pod. The PVCs
are automatically generated by `volumeClaimTemplates` field in the
StatefulSet, while the PVs have to be created manually in the
`journa-pv.yaml` in a loop.

**3. Removed the annoying range loop from master statefulset.yaml.**
All the variable scopes now come back to normal.

**4. Templated most duplicated and contrived blocks from YAML
templates.**
Moved from in the YAML templates to in the `_helper.tpl` file.

Note: This change doesn't use PV for worker storage and off-heap
storage.

pr-link: Alluxio/alluxio#10686
change-id: cid-b86ab1130d1a37f04eab11c497f31cbce33ea340