Fix workflows when a child plan has no child tasks
There were couple of bugs in the previous implementation:

1. If a plan has no children, the parent workflow currently doesn't get
its status updated correctly because it currently relies on job workers
reporting that the tasks are done/failed.

2. If a worker submits results for multiple tasks of the same job such
that the job completes, workerHeartbeat on workflow crashes. This in
particular causes https://github.com/Alluxio/alluxio/pull/10720 to break
JobWorker heartbeats.

3. Workflows are not failed properly when job workers die.

This code fixes it by replacing the workerHeartbeat logic in workflow to
being logic in the StatusChangeCallback to check for all changes in
Status instead of some.

pr-link: Alluxio/alluxio#10712
change-id: cid-5ff25d2408252e8921a419cf8e49a2f0ac446455