Add positioned readable HDFS ufs input stream
When HDFS is remote and slow, we use the positioned read API instead of
seek with this PR.

HDFS FSDataInputStream seek may be implemented as read and discard. Over
remote HDFS, this may increase the amount of data read over the network
many folds based on the workload. Using positioned read, the amount of
data requested from the HDFS client matches the amount of data read from
the HDFS datanode.

pr-link: Alluxio/alluxio#10763
change-id: cid-e343a33f11ae2dc0b8ff64d25c09926e9fc9113d