Fix directory index after disk failure
Worker uses directory index to find a directory in a tier. When a disk
goes down we detect that and remove the dir, which mess up the index. To
avoid the confusion, this fix makes the index absolute in the tier so
removing any dir does not alter its index. As a side effect,
`getDir(index)` can return null if the corresponding directory is
removed. Evictors and Allocators have been updated to handle the null
values.

pr-link: Alluxio/alluxio#10799
change-id: cid-dafe938ffbfdf987a25e0e6ffb872c72ba25a1a4