Replace Source Code Pro with Anonymous Pro font
The reason I'm proposing replacing SCP with AP is because there is a
bug in  firefox and/or MacOS which causes this  font to render
incorrectly.

See:
- https://bugzilla.mozilla.org/show_bug.cgi?id=1534503
- https://bugzilla.mozilla.org/show_bug.cgi?id=1520157

Anonymous Pro is a bit more lightweight and has at least  been recently
updated  by its maintainers (1.44.0 was 3 months ago). SCP is definitely
more ubiquitous  and well-maintained but I would rather have a font that
works everywhere. As an added bonus, this font is *much* more
lightweight than SCP.