Support compiling with Java 11
Use maven 3.6+ with JDK 11 and the flag -Pjava11 to enable the Java 11
profile. However, maven should activate the profile by default if it
detects
your default is JDK 11

There is a long list of changes and incompatibilities between Java 8
and Java 11, only a few affected us for the most part though.

- The `jdk.tools` system dependency which is transitively depended
upon by `org.apache.hbase:hbase-annotations` no longer exists in
Java 11, so we exclude `jdk.tools:jdk.tools` throughout the codebase.
- The javax annotations are no longer a part of the JDK distribution,
but they are available as a backwards-compatible maven dependency
so we add that if we're compiling with Java 11.
- The swagger maven plugin which is used to generate the REST API
docs depends on a JAXB class which isn't a part of the JDK in Java 11
so we add that dependency for the plugin to function properly.
- The maven javadoc plugin has incompatibilities for docs in a certain
format from Java 8. Configure the plugin with `<source>8</source>`
to tell the plugin the sources use some Java 8 formats. See this
OpenJDK bug for more info:
https://bugs.openjdk.java.net/browse/JDK-8212233

Partially addresses #9856

pr-link: Alluxio/alluxio#10834
change-id: cid-f8e215c55c8ca1c8f1ae0ff102f1d777696cddd0