Update table metadata from changes in UDB
This allows the sync command to update the metadata of existing
partitions, not only new partitions. This means, if an existing
partition is updated in hive, it will be detected and updated during a
sync command.

For this functionality, table version numbers and partition version
numbers have been added. The concurrency model was also updated for
these changes. It is an optimistic concurrency control model, where
multiple updates can be concurrent, but only 1 will succeed in updating
the metadata to the next version.

pr-link: Alluxio/alluxio#10836
change-id: cid-6c29985d05a17081014183ca87dd83516c36656c