Fix meter bug to calculate throughput value at master directly
Alluxio clients are usually short-lived especially those created through
running Alluxio CLI. Alluxio throughput is recorded as
`MetricType:Meter`. `Meter.getOneMinuteRate()` can be zero when the data
is first marked since the one minute rate is a one-minute
exponentially-weighted moving average rate. When a client is shutting
down, it will report metrics the last time to leader master. Because
they are too short-lived, the reported `meter.getOneMinuteRate()` is
always zero, client metrics can be considered as lost in this scenario.

In this PR, cluster throughput values are not aggregated value of
client/worker reported meter one minute rate anymore. Instead, cluster
throughput value is calculated when requested by dividing cluster bytes
counter to the metrics uptime (the last time that the metrics store is
cleared).

Because of this change, only counter value needs to be reported to the
leading master, many logics in the leading master metrics store to store
the individual metrics of types other than counter can be removed. In
addition, when the should report counters are all zero, client
heartbeats to report metrics will not be triggered. A large amount of
client heartbeat overhead is reduced.

pr-link: Alluxio/alluxio#10877
change-id: cid-0576da32984ac74ed7aa2a8e698bee78e5358d32