Choose HDFS read API based on Client buffer size
This PR is a follow up to https://github.com/Alluxio/alluxio/pull/10763.

When the client buffer size is large ( > 2MB) and reads are guaranteed
to be somewhat sequential, the `pread` API to HDFS is not as efficient
as simple `read`. We introduce a heuristic to choose which API to use.

pr-link: Alluxio/alluxio#10845
change-id: cid-3301fd510474a9c676017bbec75ade215a3eab85