Add collectInfo command
Add Alluxio CollectInfo command set to Alluxio command line. This
command will collect information on every host in the Alluxio cluster,
tarball the information into a local tarball file, then gather all the
tarballs back to where the command is issued.

This command is designed to help troubleshooting and collect information
from Alluxio clusters. The collected information include logs, config
files and it also executes a set of commands on each of the hosts and
include outputs in the tarball.

**Added commands:**
This will run one or all checks on the local machine and write the
output into files in `targetDir`. At the end of execution a tarball will
be produced containing all the information collected, put in the
`targetDir` as well.
`bin/alluxio collectInfo --local
[all/collectConfig/collectEnv/collectLog/collectMetrics/collectAlluxioInfo]
targetDir`

This will run the corresponding `collectInfo --local` command on all
hosts in the Alluxio cluster defined in `conf/masters` and
`conf/workers` via SSH. After that the tarballs generated by each
individual `infoBundle` call will be SCPed into a local temp directory,
and then put into one final tarball in the `targetDir`.
`bin/alluxio collectInfo
[all/collectConfig/collectEnv/collectLog/collectMetrics/collectAlluxioInfo]
targetDir`

**Phase 2:**
This PR is the Phase 1 of CollectInfo feature. This PR provides basic
implementations and defines the structure.

Some features are not implemented in this PR and scheduled to a separate
Phase 2 PR. These features include the following:
1. Skipping a command when pre-existing work are found. This is because
if the cluster is large, running infoBundle can take a long time and
fail in the middle.
1. Add a `-f` option to force overwrite existing work.
1. Add a `-c` or `--components` option to copy only certain logs.
1. Add a `-h` or `--hosts` option to only invoke `collectInfo` on
certain hosts.
1. Add security check to avoid copying `conf/` files that contain
credentials.

pr-link: Alluxio/alluxio#10596
change-id: cid-0bb88ff0c34bf56b14cc4b4d3f7eabc299b332e5