Fix the MultiValueMetricsAggregator and reduce overhead
Fix the bug of `MetricsSystem.getMasterMetrics()`. It previously only
returns the metric stored in MetricsRegistry with the exact same name.
In this PR, it returns the metrics that start with the given metric name
(the metrics can have multiple tags).

Reduce the logics in update `MultiValueMetricsAggregator` values, reduce
the traverse of the whole metrics.

pr-link: Alluxio/alluxio#10904
change-id: cid-3ba119c5cb9ff668dc50ca092b15e8fadd9d3dd1