Allow superuser to stopActiveSync w/o a lock
A full scan for active sync acquires a write lock on the syncPoint which
prevents the permission check for stopSync to proceed. This PR skips the
read lock needed on the inode for the permission check if the user is a
super user. This allows us to terminate a long-running full scan early.

pr-link: Alluxio/alluxio#10919
change-id: cid-e295dd75518a932382f8c17ba422e27e6842a8fa