Generate java files from .proto definitions automatically
Instead of checking in the proto-generated java files into the code
base, allow maven to automatically run protoc to generate those java
files. This adds about 10 more seconds to the build time. To skip the
building of the proto files, add `-Dskip.protoc` to the maven command.

pr-link: Alluxio/alluxio#10754
change-id: cid-d335b8ae6ac7a372aa7bb4dccd5da2686f1419f5