Refactor CommonUtils.waitFor to use value supplier
We have many instances in our code where we use CommonUtils.waitFor to
wait for a certain condition to be equal to something but the last state
of the object is not clear when there is a Timeout. This code has a new
API that could be opted into to get such a benefit.

pr-link: Alluxio/alluxio#11001
change-id: cid-b4293b9f66e390a97806c97a530fd96f346401a9