Add getWorkerAddresses and invalidate cache when registering worker
A loading cache is used for caching worker info list results. In this
way, the overhead to construct worker info list is reduced but the
worker info list result is not up to date.

Add a getWorkerAddresses() method which returns the most up-to-date live
worker addresses.
Add an invalidate cache in registering worker to make sure
getWorkerInfoList() call catchs up with the worker changes.

pr-link: Alluxio/alluxio#11048
change-id: cid-b60ef62ccbd2a84acab07b01634520389028b005