Add a script to update versions for releases
This removes all manual components for a release manager needing to
update versions throughout the codebase. In the future, any part of the
codebase which relies on the version be up-to-date should add a
section in this script that allows the developer to update the section
automatically.

pr-link: Alluxio/alluxio#10813
change-id: cid-a7025a5f6372efe78b79dfb6184a9a24e6e23082