Create new client pool map during context reinit
When a FileSystemContext is re-initialized, any acquired clients are
expected to be released back into the same pool.

Before this change, it's possible for the following scenario to occur:
- A thread would acquire a block worker client from the pool map with
key `A`.
- The context is re-initialized. All pools in the map are closed, the
pool map is cleared
- Another thread acquires a block worker client with key `A` as well.
This creates a new pool map entry with the same key `A`.
- The original thread tries to release the client it acquired before,
and it releases back into the new blockworker client pool. The pool is
not the same as before, and an exception is thrown.


Fixes #11221

pr-link: Alluxio/alluxio#11218
change-id: cid-a5c4e568f3e18cdf0432733b33448681abc86253