Fix LockPoolTest
I'm not sure how this test didn't fail more consistently before.

Previously, the test would insert values [0,15] into the lock pool.
Then, insert 16 in order to trigger an eviction. Next, it would finally
attempt to insert [16,23]. Which should go up to the high watermark.
However, sometimes the lock pool would evict entries [0, 8] (instead
of [0,7]) leaving [9, 16], and attempting to insert [16, 23]. [9,23] is
only 15 total entries, when it expects 16. I make the test now insert
[17,24]. It inserts the same amount of keys. It's just that the key
range is shifted, so that there is no key overlap.

The reason is because the evictor thread calculates numToEvict as
pool size - low watermark. After inserting the 17th item, the
numToEvict is 17 - 8 -> 9. For some reason before my changes,
this was being calculated as 8. As if the evictor was triggered prior
to the key being added.