Split sync executor size from sync concurrency
This can help limit the impact of a single sync job on the
performance of Alluxio as a whole. This is especially
important if more than one path is being synced at a time.
It allows for a limited level of concurrency but still lets
us share a single thread pool for executing sync tasks.