Keep state lock for duration of journal context
The state change lock should be held for the duration of the RPC, so
that when the pause lock is held, there are no outstanding RPCs. This
change makes the state change lock to be held during the whole journal
context, so that when backups pause the state, it is guaranteed there
are no state changes or journal changes in progress.

Fixes #11259

pr-link: Alluxio/alluxio#11255
change-id: cid-8a047cbe4d5cf2a15b9d72bd1591f6ac71c320f6