Improve performance of sasl authentication
Before this PR, the ImpersonationAuthenticator class could effectively
bottleneck the amount of concurrent channels that a worker could handle.

The `<init>` for the ImpersonationAuthenticator class required iterating
over
all keys within an Alluxio configuration, not only doing key lookups,
but also
performing regex matching. For some keys this could be especially
expensive
(i.e, default supplier which calls
Runtime.getRuntime().availableProcessors()).
On top of that, performing regex matching is also quite expensive for a
few
hundred keys.

After taking a quick peek at the code it seems that it's entirely
unnecessary to
instantiate the impersonation authenticator for every new handler that
we
create. Instead, it can just be a field within the
DefaultAuthenticationServer.
Then, it only ever needs to be instantiated once.

pr-link: Alluxio/alluxio#11293
change-id: cid-96438acee91197e2e9a380494848db481c458660