Stop tracking connections in copycat client/server
Many connections can be created by the transport layer, but it is rare
for copycat to close the client. Tracking the created connections is
unnecessary because copycat closes them before closing the clients
anyways. This prevents us from holding on to the connection references
for extended periods of time, which then causes behavior resembling a
memory leak.

Additionally we bump the copycat version to 1.2.15 to avoid cycling
connections. 1.2.15 introduces a changes to the KeepAliveResponse serde
to respect unresolved addresses.

Fixes #11355

pr-link: Alluxio/alluxio#11354
change-id: cid-8e00a34b6fe9f54d039c7735da0f57328e4f18d7