Improve PR build by inspecting PR diff
we currently use this jenkins plugin to run our PR build checks:
https://plugins.jenkins.io/ghprb/

the plugin defines an environment variable for the target branch that
the PR is going to merge into. that means we can run `git --no-pager
diff origin/master --name-only` to get a list of files that were changed
in the PR. note this change also required a few tweaks to the jenkins
job, most importantly the addition of pulling more refs so i can
reference the ref marked as origin/master
since we have two checks, one for the maven build and one for docs, i
defined two booleans that get set to true if a corresponding file is
found in the diff. if this target branch variable is not available, it
defaults to running all checks.
if there are more files or folders that we can incorporate into the diff
check, we can waste less cycles running irrelevant checks

pr-link: Alluxio/alluxio#11376
change-id: cid-4c42a1601297fe69aa66b7cb790c2b7a74621964