Improve HdfsPositionedReadUnderFileInputStream with a heuristic
The heuristic works as follows :
1. it starts with a pread, because a lot of times the client only does a
single large read, and we want to benefit from pread in that case.
2. if there are certain number of sequential reads in a row, we switch
to buffered read mode. In this mode, preads are serviced via pread
calls, and read calls are services by read calls. A sequential read is
defined as a read that is within a movement limit of the previous read.
This is to guard against workloads such as read, skip(2), read, skip(3)
etc.
3. any non-sequential read, we change back to pread mode.

pr-link: Alluxio/alluxio#11400
change-id: cid-482d194fd551a9e80892682a90cc5e82290e284a