Improve metadata sync operations
This PR improves the "syncMetadata" operation used throughout the
DefaultFileSystemMaster.

This approach incrementally locks and syncs the paths in the tree
instead of taking a write lock on the root of the tree. The
"loadMetadata" step of syncing is now contained within the
"syncInodeMetadata" function.

It is implemented with a BFS. The previous implementation used DFS which
could have been prone to stack overflows if a user created a deep enough
file hierarchy.

Additionally, this implementation makes the minimum amount of RPC calls
required to the UFS. UfsStatuses are re-used as much as possible, and
discarded once they are no longer required.

pr-link: Alluxio/alluxio#11204
change-id: cid-e6a3d62e25ee31cefe1156f7f9f32d85ae5a8c64