Make Ozone as a under file system of Alluxio
This PR introduces a new top-level UFS module for compiling an
Ozone-compatible UFS. The Ozone client is able to use the Hadoop
FileSystem interface, so we don't need to write any new logic to support
Ozone.

The reason we include this as a top-level UFS as opposed to including
the ozone libraries directly in the HDFS UFS is to gain more control
over the supported ozone version as well as reduce the size of the
release tarball. Otherwise, we would maintain many copies of the same
Ozone jars.

Resolves #11390

pr-link: Alluxio/alluxio#11396
change-id: cid-06387b11a6bd2ce0694cf33bad1f09f76a396c2c