Add readonly check while delete recursively
When I use `rmr` to delete an alluxio path with a descendant which is a
readonly mount point, alluxio will delete all the descendants no matter
if it is a readonly mount point, this is extremely dangerous. So it's
important to add readonly check while delete recursively.

**Use-case:**
```
./bin/alluxio fs mkdir /tmp
./bin/alluxio fs mount --readonly /tmp/mnt hdfs://example/tmp/mnt
```
UserA with all the permission did:
```
./bin/alluxio fs rmr /tmp/mnt
A write operation on /tmp/mnt under a readonly mount point
hdfs://example/tmp/mnt is not allowed
```
UserA with all the permission did:
```
./bin/alluxio fs rmr /tmp
# Everything under hdfs://example/tmp/mnt gone
```
As a result, `hdfs://example/tmp/mnt` was unexpectedly deleted

pr-link: Alluxio/alluxio#11513
change-id: cid-70fce19c6d2e72a080919597e028aa1a54c6ef74